Organize Task component
